Fixes to previous commit

+++ GraphicsTestFramework_Core/CoreUtility/String/StringToUI.cs
@@ -7,13 +7,12 @@ namespace CoreUtilities
 	
 	public static class StringToUI
 	{
-
-		public static string Convert (string input)
+        public static string Convert (string input)
 		{
 			char[] array = input.ToCharArray ();
 			string output = "";
 			for (int i = 0; i < array.Length; i++) {
-				output += array [i];
+				output += (int)array [i];
 			}
 			return output;
 		}
